Backend/requirements.txt
django
django-heroku
gunicorn
djangorestframework
pyexcel-xlsx